+++ CHANGELOG.md
@@ -5,11 +5,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- DFSNamespaceRoot
+  - Added support for setting the state of the namespace root (Status)
+- DFSNamespaceFolder
+  - Added support for setting the state of the namespace folder (Status)
+
 ## [5.0.1] - 2023-09-07
 
 - DFSDsc
   - Removed the pipeline files `build.psd1` as it is no longer needed.
-  
+
 ## [5.0.0] - 2023-09-01
 
 ### Changed