+++ CHANGELOG.md
@@ -5,11 +5,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- DFSNamespaceServerConfiguration
+  - Added support for setting EnableSiteCostedReferrals, EnableInsiteReferrals and PreferLogonDC on a DFS namespace server
+
 ## [5.0.1] - 2023-09-07
 
 - DFSDsc
   - Removed the pipeline files `build.psd1` as it is no longer needed.
-  
+
 ## [5.0.0] - 2023-09-01
 
 ### Changed