+++ CHANGELOG.md
@@ -5,6 +5,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Updated out of date README.MD.
+
 ## [5.0.1] - 2023-09-07
 
 - DFSDsc