+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Added support for setting the state of the namespace root (State)
 - DFSNamespaceFolder
   - Added support for setting the state of the namespace folder (State)
+- Updated out of date README.MD.
 
 ## [5.0.1] - 2023-09-07