+++ CHANGELOG.md
@@ -5,13 +5,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
 - DFSNamespaceRoot
   - Added support for setting the state of the namespace root (State)
 - DFSNamespaceFolder
   - Added support for setting the state of the namespace folder (State)
+ 
+### Fixed
+
 - Updated out of date README.MD.
 
 ## [5.0.1] - 2023-09-07
+
+### Fixed
 
 - DFSDsc
   - Removed the pipeline files `build.psd1` as it is no longer needed.